Rename deprecation module
